requirements_fixed.txt
# requirements_fixed.txt
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.3.0
mss==9.0.1
pytesseract==0.3.10
pyyaml==6.0.1
pyautogui==0.9.54